ren mit Exception falls Pfad schon vorhanden

+++ commandercmd.console/Commands/renCommand.cs
@@ -16,23 +16,30 @@ namespace commandercmd.console.Commands
         public override void Execute()
         {
             
-            Console.WriteLine("Hoi Stephhhhhhh");
+           
             Console.WriteLine(this.Parameter);
             String startpath = this.Parameter;
             String[] endpath = startpath.Split(' ');
 
-            File.Move(endpath[0], endpath[1]);
+            if (!File.Exists(endpath[0]) && !File.Exists(endpath[1]))
+            {
+                Console.WriteLine(" Ihr Pfad gibt es bereits");
+            }
+            
+            else {
+                File.Move(endpath[0], endpath[1]);  
+                Console.Write("Ihre Datei wurde verschoben");
+            
+            }
+            
 
-            //Console.WriteLine(endpath[0], endpath[1]);
+            
 
-            //if(this.Parameter!= null)
-            //{
-             //   Program.shell.Exit();
-            //}
+            
             
 
 
             
-        }
+        
     }
 }
